gocos: use double for clock values to avoid implicit cast to float

+++ src/gocos/golog_adapter.cpp
@@ -38,10 +38,10 @@ operator<(const GologLocation &first, const GologLocation &second)
 }
 
 namespace details {
-std::map<std::string, float>
+std::map<std::string, double>
 get_clock_values(const ClockSetValuation &clock_valuations)
 {
-	std::map<std::string, float> res;
+	std::map<std::string, double> res;
 	for (const auto &clock : clock_valuations) {
 		res[clock.first] = clock.second.get_valuation();
 	}
